Modified sleep prep, payload format changes, sensor range validation, added antenna voltage switch

+++ LoRaWAN-NZ865-NZ32-inAir9B-Ultrasonic/app/bcd.h
@@ -1,6 +1,6 @@
 #ifndef _BCD_H
 #define _BCD_H
 
-void formatPayload(uint8_t *outbuf, uint16_t range, int16_t temperature, uint8_t volts );
+void formatPayload(uint8_t *outbuf, uint16_t range, int16_t temperature, uint16_t volts );
 
 #endif // _BCD_H
